requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.12.4
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
